(feat) add first-page carousel
