+++ src/components/newsletter-swiper/index.scss
@@ -1,0 +1,62 @@
+.swiper-button-next:active,
+.swiper-button-prev:active {
+    color: #00b1fd;
+}
+
+.swiper-pagination-bullet {
+    background: #d9d9d9;
+    margin: 0 0.5rem !important;
+}
+
+.swiper-pagination-bullet-active {
+    background: var(--ifm-color-primary);
+}
+
+.firstPageSwiper {
+    .swiper-pagination-bullets.swiper-pagination-horizontal {
+        left: 155px;
+        top: 150px;
+        width: auto;
+        .swiper-pagination-bullet {
+            background-color: rgba(255, 255, 255, 0.5);
+            opacity: 100%;
+            height: 0.25rem;
+            width: 0.25rem;
+        }
+        .swiper-pagination-bullet-active {
+            width: 2.5rem;
+            border-radius: 0.375rem;
+            background: linear-gradient(to right, white var(--progress-count), rgba(255, 255, 255, 0.5) 0%) !important;
+        }
+    }
+}
+
+@media (max-width: 1280px) {
+    .firstPageSwiper {
+        .swiper-pagination-bullets.swiper-pagination-horizontal {
+            top: 35%;
+            left: 50%;
+            transform: translateX(-50%);
+        }
+    }
+}
+
+@media (max-width: 1024px) {
+    .firstPageSwiper {
+        .swiper-pagination-bullets.swiper-pagination-horizontal {
+            top: 38%;
+            left: 50%;
+            transform: translateX(-50%);
+        }
+    }
+}
+
+@media (max-width: 640px) {
+    .firstPageSwiper {
+        .swiper-pagination-bullets.swiper-pagination-horizontal {
+            top: 33%;
+            left: 50%;
+            transform: translateX(-50%);
+        }
+    }
+}
